Update tests
- Remove mock Repository from tests. For simplicity, I will test repo alongside service and validator tests. This will ensure that data is actually being added/retrieved/updated/deleted

+++ src/repository/Repository.ts
@@ -19,6 +19,9 @@ export default class Repository<T extends Model> {
   }
 
   public delete(key: number) {
+    if (!this.storage.has(key)) {
+      throw new Error(`Item with key ${key} not found`);
+    }
     this.storage.delete(key);
   }
 }
